adds ws base-client, WIP (affects #102)
