+++ websocket/cmd/event.go
