+++ websocket/disagnose_false.go
@@ -4,5 +4,5 @@ package websocket
 
 const SaveIncomingPackets = false
 
-func saveOutgoingPacket(c *Client, packet *clientPacket)                {}
-func saveIncomingPacker(c *Client, event *discordPacket, packet []byte) {}
+func saveOutgoingPacket(c *baseClient, packet *clientPacket)                {}
+func saveIncomingPacker(c *baseClient, event *discordPacket, packet []byte) {}
