+++ websocket/opcode/voice.go
@@ -1,0 +1,18 @@
+package opcode
+
+const (
+	VoiceIdentify uint = iota
+	VoiceSelectProtocol
+	VoiceReady
+	VoiceHeartbeat
+	VoiceSessionDescription
+	VoiceSpeaking
+	VoiceHeartbeatAck
+	VoiceResume
+	VoiceHello
+	VoiceResumed
+	_ // unused
+	_ // unused
+	_ // unused
+	VoiceClientDisconnect
+)
